Override delegateEvents instead of initialize
The initialize() method of Backbone.View is a no-op that is meant to be
overridden without the need to call super. Using this plugin means that
every class that extends Backbone.View now has to invoke super in its
initialize, where it never did before.

Overriding delegateEvents/undelegateEvents rather than initialize/remove
fixes this annoyance, and also seems close to the spirit of those
methods (i.e., if you manually invoked undelegateEvents on a view, you'd
probably want it to undelegate keyboard events too.)

Add back the unbindKeyboardEvents on remove()

The undelegateEvents method doesn't get invoked on remove(), so we
actually need the hook in both places.